+++ ui.frontend/src/components/Results/ResultsExporter/index.ts
@@ -1,2 +1,2 @@
 export * from './ResultsExporter';
-export * from './ResultsExporterModal';
+export * from './ResultsExporterDialog';