+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 
 ## Content Wizard 1.1.4 - July TBD, 2024
 
-- Refactored the Results exporting logic to use a Modal with a configuration form, that allows the user to specify file type, file name, and if the file name should include a timestamp.
-- Converted Modal Close buttons into IconButtons from FABs, and cleaned up modal styles.
+- Refactored the Results exporting logic to use a Dialog with a configuration form, that allows the user to specify file type, file name, and if the file name should include a timestamp.
+- Converted Dialog Close buttons into IconButtons from FABs, and cleaned up dialog styles.
+- Implemented the Package Builder Dialog, which allows users to Create and Build a Content Package for CRX PackageManager, based on the results.
 
 ## Content Wizard 1.1.3 - July 18, 2024